requirements.txt
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.1
mongoengine==0.20.0
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
pip==20.2.3
pymongo==3.11.0
python-dotenv==0.14.0
pytz==2020.1
setuptools==41.2.0
six==1.15.0
sqlparse==0.3.1
wheel==0.33.1
certifi==2020.6.20
idna==2.10
requests==2.24.0
urllib3==1.25.10
chardet==3.0.4